# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Modular monolith backend (Go + Chi). HTTP handlers live in domain packages (`auth/handler`, `events/handler`, `ticketing/handler`, …) under `internal/app` at base path **`/api/v1`** (отдельного пакета `internal/api/v1` в репозитории нет).\n**Ошибки:** тело `{ \"error\": { \"code\": \"string\", \"message\": \"string\" } }`. Типичные HTTP-коды: **401 Unauthorized** — нет/невалидный JWT (`missing_authorization`, `invalid_token`, `invalid_credentials`, …); **403 Forbidden** — RBAC (`forbidden`, `organizer_request_forbidden`, …); **409 Conflict** — билеты и бизнес-правила (`already_registered`, `capacity_full`, `event_not_approved`, …). Полный список `error.code` — в README.\nДаты событий: **RFC3339**, например `2026-01-01T10:00:00Z`. Ответ `POST /tickets/register`: поля **`qr_hash_hex`** и **`qr_png_base64`**.",
        "title": "Student Event Ticketing Platform API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/notifications/send-email": {
            "post": {
                "description": "No JWT required in current build. May return 501 if sending is not wired.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Enqueue outbound email (foundation)",
                "parameters": [
                    {
                        "description": "Email payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SendEmailRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted — enqueued"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.SendEmailRequestDTO": {
            "type": "object",
            "required": [
                "body",
                "title",
                "to"
            ],
            "properties": {
                "body": {
                    "type": "string",
                    "maxLength": 5000,
                    "minLength": 1
                },
                "title": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 3
                },
                "to": {
                    "type": "string"
                }
            }
        },
...
        "httpx.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "httpx.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/httpx.ErrorBody"
                }
            }
        },
...
    }
}